--- src/main/java/ru/yandex/practicum/filmorate/model/IUserService.java
@@ -1,28 +1,0 @@
-package ru.yandex.practicum.filmorate.model;
-
-import ru.yandex.practicum.filmorate.exception.NotFoundException;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
-
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-
-public interface IUserService {
-    public Collection<User> getAll();
-
-    public Map<Long, User> getUsers();
-
-    public User getUserById(Long userId) throws NotFoundException;
-
-    public User create(User user);
-
-    public User update(User user) throws ValidationException;
-
-    public User addFriend(Long userId, Long friendId) throws ValidationException;
-
-    public void deleteFriend(Long userId, Long friendId) throws ValidationException;
-
-    public List<User> getUserFriends(Long userId) throws NotFoundException;
-
-    public List<User> getUserCommonFriends(Long userId, Long otherId) throws ValidationException;
-}
