+++ src/test/java/ru/yandex/practicum/filmorate/controller/FilmControllerTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.service.FilmService;
-import ru.yandex.practicum.filmorate.service.UserService;
+import ru.yandex.practicum.filmorate.service.FilmServiceImpl;
+import ru.yandex.practicum.filmorate.service.UserServiceImpl;
 import ru.yandex.practicum.filmorate.storage.InMemoryFilmStorage;
 import ru.yandex.practicum.filmorate.storage.InMemoryUserStorage;
 
@@ -17,7 +17,7 @@ public class FilmControllerTest {
 
     @BeforeEach
     void setUp() {
-        FilmController filmController = new FilmController(new FilmService(new InMemoryFilmStorage(), new UserService(new InMemoryUserStorage())));
+        FilmController filmController = new FilmController(new FilmServiceImpl(new InMemoryFilmStorage(), new UserServiceImpl(new InMemoryUserStorage())));
     }
 
     @Test()
